backend/fastapi/requirements.txt
fastapi[all]
uvicorn[standard]
gunicorn
motor[srv]
python-dotenv
SQLAlchemy
alembic
opentelemetry-sdk 
opentelemetry-exporter-otlp 
opentelemetry-instrumentation-fastapi 
opentelemetry-instrumentation-requests 
opentelemetry-instrumentation-logging 
protobuf
boto3
psycopg2-binary
python-multipart
python-jose
asyncpg
aioboto3
passlib[bcrypt]
aiosmtplib
redis